% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi3.R
\name{bmi3}
\alias{bmi3}
\title{Create BMI Groups}
\usage{
bmi3(x)
}
\description{
More detailed description.
}
\details{
@param x Numeric vector.

 @return Factor variable

 @examples
 bmi.vals <- rnorm(n = 20, mean = 25, sd = 3)
 bmi3(bmi.vals)

 @export
}
