% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package 1.R
\docType{package}
\name{package1}
\alias{MMPackage}
\alias{MMPackage-package}
\alias{package1}
\title{BMI Package}
\description{
Contains one function for creating BMI groups.
}
\author{
Dilara Tuncyürek \email{d.tuncyuerek@campus.lmu.de}
}
